# shared/api/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  piemdm_internal_model.Application:
    properties:
      appId:
        description: AppId
        maxLength: 64
        type: string
      appSecret:
        description: AppSecret
        maxLength: 128
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 简介
        maxLength: 255
        type: string
      id:
        type: integer
      ip:
        description: 系统来源IP
        type: string
      name:
        description: 系统名称
        maxLength: 128
        type: string
      status:
        description: 状态：Normal 正常 Frozen 已冻结 Deleted 已删除
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
host: localhost:8787
info:
  contact:
    email: support@piemdm.com
    name: API Support
  description: PieMDM Master Data Management System API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PieMDM API
  version: "1.0"
paths:
  /admin/applications:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 15
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 应用ID
        in: query
        name: appId
        type: string
      - description: 开始日期
        in: query
        name: startDate
        type: string
      - description: 结束日期
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.Application'
            type: array
      summary: 获取应用列表
      tags:
      - 应用管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建应用请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.Application'
      summary: 创建应用
      tags:
      - 应用管理
  /admin/applications/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除应用
      tags:
      - 应用管理
    get:
      consumes:
      - application/json
      parameters:
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.Application'
      summary: 获取应用详情
      tags:
      - 应用管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新应用请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新应用
      tags:
      - 应用管理
  /admin/applications/batch:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量删除应用
      tags:
      - 应用管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量更新应用状态
      tags:
      - 应用管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"